fix: firmware.c: wrong new function usage
GetFirmwareType function is usable on Windows version larger than 8.
Directly using it in a program can make the program not usable on
system whose version lesser than 8, for Windows will find no such
symbol in the libraries linked.

Use GetProcAddress to dynamically get the handle of GetFirmwareType,
and then invoke it, to solve this problem.

Signed-off-by: Junde Yhi <lmy441900@gmail.com>